% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_sample.R
\name{joint_sample}
\alias{joint_sample}
\title{purpose of function: Jointly draw parameters from asymptotic distribution of ML estimates.
Samples can be used to empirically construct CI's for the RFV or th survival function.
Note that CIs wont be supplied by function, only samples will be drawn.
Consider to set seed beforehand!}
\usage{
joint_sample(
  n_samp = 1e+05,
  marginal_par,
  cond_par = NULL,
  H,
  exclude = FALSE,
  H.singular = "standard"
)
}
\arguments{
\item{n_samp}{number of samples to be drawn from multivariate normal}

\item{marginal_par}{all parameters from marginal$par, i.e. from optimization, on optimzed scale}

\item{cond_par}{same for conditional model parameters (beta, hazard) (irrelevant for current status data)}

\item{H}{Hessian. Order of parameters has to be in line with c(marginal_par,cond_par).  no further parameters allowed!}

\item{exclude}{specify positions of parameters that should be excluded from being samples}

\item{H.singular}{Handling of hessian if it is singular.}
}
\value{
list of samples and input variance of multivariate normal form which samples were drawn.
}
\description{
purpose of function: Jointly draw parameters from asymptotic distribution of ML estimates.
Samples can be used to empirically construct CI's for the RFV or th survival function.
Note that CIs wont be supplied by function, only samples will be drawn.
Consider to set seed beforehand!
}
