% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{LRT}
\alias{LRT}
\title{LRT}
\usage{
LRT(l_h1, l_h0, df)
}
\arguments{
\item{l_h1}{loglihood value of larger model}

\item{l_h0}{loglihood value of nested model}

\item{df}{degrees of fredom}
}
\value{
numeric vector with LRT, df, -value
}
\description{
LRT
}
