% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_AIC.R
\name{LRT_AIC}
\alias{LRT_AIC}
\title{compares either a full (H1) and a nested (H0) model or, alternatively,
compares alpha_gamma vs gamma and alpha_gamma vs alpha.
list elements must be names in the nlm sense ($estimate, $minimum)}
\usage{
LRT_AIC(alpha_gamma = NULL, alpha = NULL, gamma = NULL, H1 = NULL, H0 = NULL)
}
\arguments{
\item{alpha_gamma}{AF model (optional), alternatively H1, H0 might be specified}

\item{alpha}{Poisson restricted model (optional), alternatively H1, H0 might be specified}

\item{gamma}{gamma frailty model (optional), alternatively H1, H0 might be specified}

\item{H1}{Some model (optional2), alternatively alpha_gamma, alpha, gamma might be specified}

\item{H0}{Model nested in H1 model (optional2), alternatively alpha_gamma, alpha, gamma might be specified}
}
\value{
LRT, AIC, Deviance, Deviance-Test
}
\description{
compares either a full (H1) and a nested (H0) model or, alternatively,
compares alpha_gamma vs gamma and alpha_gamma vs alpha.
list elements must be names in the nlm sense ($estimate, $minimum)
}
