% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_surv.R
\name{ph.surv}
\alias{ph.surv}
\title{Fit univariate or multivariate time-to-event models with Addams Family or PVF frailty or no frailty
with piecewise constant, Breslow or generalized gamma baseline hazard and proportional hazard factors.}
\usage{
ph.surv(
  data,
  X = NULL,
  strata = NULL,
  y,
  d,
  frailty = NULL,
  ID = NULL,
  tr = NULL,
  weights = NULL,
  start_dist = NA,
  start_lnlam = NA,
  start_beta = NA,
  current.status = FALSE,
  Random.Slope = NA,
  slope.model = "free-Willy",
  binom.n = 0,
  start_slope = NA,
  start_pi = NA,
  baseline = list("piecewise", "Breslow", NULL, NULL),
  type,
  ci_level = 0.95,
  overdisp.cs = FALSE,
  scaleX = TRUE,
  centerX = TRUE,
  breaks = NULL,
  iterlim = 1000,
  reltol = 1e-10,
  converge = 1e-06,
  method = "BFGS",
  c.method = method,
  m.method = method,
  Richardson = FALSE,
  R.options = list(),
  m.options = list(),
  c.options = list(),
  multi.num = FALSE,
  cs.options = list(EM.cs = FALSE, univariate = TRUE, uni.model = NA, uni.initial = TRUE),
  frail.thresh = 1e-06,
  H.singular = "standard",
  scale.marginal = -1,
  scale.cond = -1,
  z = NA,
  v = NA,
  initial.z = TRUE,
  prnt = FALSE
)
}
\arguments{
\item{data}{data.frame. Only numeric values or logical or factors are allowed (latter will be coerced to numeric).
Note that multilevel factors as proportional hazard covariates are not yet supported}

\item{X}{vectors of covariate (column) names referring to data}

\item{strata}{string that refers to stratum variable (column) in data. Can be mutli-level.
Null if no stratified hazard}

\item{y}{name of time variable in data}

\item{d}{name of status indicator in data (should be 1 for event, 0 for censored)}

\item{frailty}{name stratum variable for frailty distrbution (models for frailty distribution).
NULL if frailty distribution is without model or no frailty mode is fitted}

\item{ID}{name of ID variable to indicate clusters}

\item{tr}{name of left-truncation time variable. Only used for right-censored data! can be 0 or NULL in case of no truncation.}

\item{weights}{name of weight variable. weighting factors twhitch which individual likelihood contribution is multiplied with}

\item{start_dist}{starting values of frailty parameters}

\item{start_lnlam}{starting values of hazard parameters (note that in some cases log hazard parameters are used, e.g. piecewise hazard)}

\item{start_beta}{starting values of PH coeffiecients}

\item{current.status}{logical. If TRUE current status data models will be fitted, right censored data if FALSE}

\item{Random.Slope}{logical (if random slopes should be imposed. experimental. Better ignore)}

\item{slope.model}{model for random slope. Was experimental. Should be ignored}

\item{binom.n}{if aB or B+p is chosen in type the this is the number of RC chosen.
Might also be number of RC for binomial random slopes (which is experimental, better don't model random slopes and better stratify frailty distribution by option frailty)}

\item{start_slope}{starting values ranodm slopes}

\item{start_pi}{starting values random slopes}

\item{baseline}{list up to length 4. Specifies the baseline hazard model.
Default is baseline = list('piecewise','Breslow',NULL,NULL)
The first element of the list can be a vector of length strata specifying the baseline hazards.
Options are piecewise, GG, and its special cases
c('gamma', 'gamma^-1', 'Weibull', 'Weibull^-1', 'exponential', 'exponential^-1', 'ammag', 'ammag^-1', 'lognormal', 'location', 'N/2', 'GGpos', 'GGinv' ).
Bseline hazard distibutions can differ for strata.
The second element of the list specifies handling of piecewise hazards. Can be Intervall or Breslow. In case of intervall the third element of the list is the step size of intervals.
Alternatively breaks.piece can be specified for customized time intervals for dsitnct hazard parameters.
Fourth elemnt of list specifies a groups for which all GG parameters except for the location parameter are shared.
Only relevant if location is chosen as baseline hazard (first element of list).
Groups are numbered strata, can be more than one group,
i.e. fourth element can be a list.
#' @param type c('alpha_gamma', 'alpha', 'gamma','PVF', 'cP', 'IG', 'H')
Frailty distibution (first is Addams, second scaled Poisson).
Can be of length(unique(data$frailty)), i.e. differnet distribution for varying strata can be chosen.}

\item{ci_level}{level on CIs}

\item{overdisp.cs}{logical. If true overdispersion model will be estimated (Compound multinomial Dirichilet).
Only available for current status data.}

\item{scaleX}{logical. Should covariates be scaled}

\item{centerX}{logical, should covariate be centered}

\item{breaks}{list of break points for distinct hazard parameters for piecewise hazard}

\item{iterlim}{maximum number of iterations}

\item{converge}{convergence threshold. Note that convergence criterion is abolute not relative increment in loglihood}

\item{method}{optimisazion mehtod. Essentiall everything that R hast to offer
can be chosen (optimize, every method of optim, nlm, nlminb, DEoptim, constrOptim, genoud...)}

\item{c.method}{same as method but for conditional model}

\item{m.method}{same as above butt for marignal model}

\item{Richardson}{logical. Indicates whther Richardson derivative as implementd in numDeriv should be used}

\item{R.options}{list of options for numDeriv partameters.
Exact naing is a must. Parameters that are not set here are default values}

\item{m.options}{list of options for the optimizer of maringal model.
Names must be exact the same as the options of chosen optimzation routines. All options that are not set will be set to default values.}

\item{c.options}{same as m.options for conditional model}

\item{multi.num}{whther numerical derivatives should be used.
For bivariate tight censored models derivatives and (partially) hessians are analytically derived}

\item{cs.options}{options for current statuts data optimization.
List list(EM.cs, univariate,uni.model=NA,uni.initial), First states whether EM algorithm should be used (set to FALSE, EM is horrible in this case), second whether univariate model should be also fitted, third can be the univariate model such that it does not have to be omputed more than once if you fit more than one model, fourth whether hazard parameters of maringal model should be initialized with univariate model (my impression is that this often works better for current status data).}

\item{frail.thresh}{threshold for borderline cases, i.e. when should we assume that the AF model is the gamma, Poisson etc model. Might be important for optimization purposes as this changes how survival function is computed in current optimization iteration}

\item{H.singular}{treatment if Hessian is singualr. Can be gcholesky, ginverse, standard.}

\item{scale.marginal}{scaling of maringal loglihood. Actually makes no sense.
Can better be handled via sspecifiying options for optimization routines.}

\item{scale.cond}{same as above}

\item{z}{initinal values of the frailties}

\item{v}{intial values of the random slopes}

\item{initial.z}{logical indicator. Specifies whther frailties should be intialized with intial frailty distribution paramers. If FALSE frailties are 1 in first iteration (only relevant for right censored data)}

\item{prnt}{logical. indicates whther current iteration should be printed.}
}
\value{
list of parameter estimates, choice of distribution and further estimates for plotting and prediction purposes
}
\description{
Fit univariate or multivariate time-to-event models with Addams Family or PVF frailty or no frailty
with piecewise constant, Breslow or generalized gamma baseline hazard and proportional hazard factors.
}
\examples{
'Please see R markdown script that is also uploaded on git.
Documentaion and vignette is work in progress...'
}
