% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_S.R
\name{predict.S_s}
\alias{predict.S_s}
\title{Computes Survival function given survival status of cluster comrades.}
\usage{
\method{predict}{S_s}(
  object,
  newdata_t,
  newdata_s = 0,
  Lam_ij,
  time = 1,
  status = 2,
  frailty.strata,
  ID
)
}
\arguments{
\item{object}{Fitted object from ph.surv}

\item{newdata_t}{data frame with surivavl times for which survival function should be calculated (not the conditions!)}

\item{newdata_s}{data frame with conditions, i.e. known survival times up to s.}

\item{Lam_ij}{matrix with covariate weighted cumulative baseline hazard at time t in first column and at time f condition in second column (see predict.haz)}

\item{time}{name or position of time variable}

\item{status}{name or position of event indicator}

\item{frailty.strata}{name or position of frailty stratum}

\item{ID}{name or position of ID variable}
}
\value{
predicted conditional survival functions as matrix
}
\description{
Computes Survival function given survival status of cluster comrades.
}
