% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowall.R
\name{rowall}
\alias{rowall}
\title{row-specific maximum}
\usage{
rowall(x, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{...}{further arguments passed to max (na.rm)}
}
\value{
logical vector of length dim(x)\link{1}
}
\description{
row-specific maximum
}
\examples{
rowall(cbind(c(1,2,3),c(1,1,10))==1)
rowall(cbind(c(NA,2,3),c(1,1,10)==1), na.rm = TRUE)
}
