% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowany.R
\name{rowany}
\alias{rowany}
\title{row-specific maximum}
\usage{
rowany(x, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{...}{further arguments passed to max (na.rm)}
}
\value{
logical vector of length dim(x)\link{1}
}
\description{
row-specific maximum
}
\examples{
rowany(cbind(c(1,2,3),c(1,1,10))==1)
rowany(cbind(c(NA,2,3),c(1,1,10)==1), na.rm = TRUE)
}
