% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmax.R
\name{rowmax}
\alias{rowmax}
\title{row-specific maximum}
\usage{
rowmax(x, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{...}{further arguments passed to max (na.rm)}
}
\value{
numeric vector of length dim(x)\link{1} with row-secific maxima
}
\description{
row-specific maximum
}
\examples{
rowmax(cbind(c(1,2,3),c(1,1,10)))
rowmax(cbind(c(NA,2,3),c(1,1,10)), na.rm = TRUE)
}
