% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_dist.R
\name{se.dist}
\alias{se.dist}
\title{Computes confidence intervals for for RC frailty values and conditional hazard ratios}
\usage{
se.dist(
  zstar = 0:10,
  model.par,
  mu = 1,
  Sigma,
  type = "alpha_gamma",
  num.a = 1,
  num.g = 1,
  Kstar = 0,
  mean.stratum = FALSE,
  RE.stratum = row.names(model.par),
  CI = 0.95,
  dist.par
)
}
\arguments{
\item{dist.par}{}
}
\description{
Computes confidence intervals for for RC frailty values and conditional hazard ratios
}
